requirements.txt
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
greenlet==1.1.2
iniconfig==1.1.1
packaging==21.3
playwright==1.17.2
pluggy==1.0.0
psycopg==3.0.8
psycopg-binary==3.0.8
psycopg-pool==3.0.3
py==1.11.0
pydantic==1.8.2
pyee==8.2.2
pyparsing==3.0.6
pytest==6.2.5
python-dotenv==0.19.2
soupsieve==2.3.1
toml==0.10.2
typing_extensions==4.0.1
websockets==10.1
